Allow key events to traverse the TopLevel hierarchy
It's quite common for key events to flow up the containment
hierarchy in UI frameworks. This could simplify management
of global hotkeys that can be placed in a top-level "App"
and have all views shown subsequently just "inherit" that
behavior as long as they let the key event to flow.